# swagger.yml
openapi: '3.0.2'
info:
  title: Masp Indexer
  version: '1.0'
servers:
  - url: https://localhost:5000/api/v1
paths:
  /health:
    get:
      responses:
        '200':
          description: OK
  /commitment-tree:
    get:
      parameters:
        - in: query
          name: block_height
          schema:
            type: integer
      responses:
        '200':
          description: The commitment tree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeResponse'
  /notes-map:
    get:
      parameters:
        - in: query
          name: from-height
          required: true
          schema:
            type: integer
        - in: query
          name: to-height
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The notes map between two block heights
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeResponse'
  /witness-map:
    get:
      parameters:
        - in: query
          name: block_height
          schema:
            type: integer
      responses:
        '200':
          description: The witness map of a specific block height
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeResponse'
  /tx:
    get:
      parameters:
        - in: query
          name: from-height
          required: true
          schema:
            type: integer
        - in: query
          name: to-height
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The masp transactions between two block height
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TreeResponse'

components:
  schemas:
    TreeResponse:
      type: object
      properties:
        commitment_tree:
          type: string
          format: byte
          description: The commitment tree bytes.
        block_height:
          type: integer
          description: The block height of the commitment tree
    WitnessMapResponse:
      type: object
      properties:
        witnesses:
          type: array
          items:
            type: object
            properties:
              bytes:
                type: string
                format: byte
                description: The witness bytes.
              index:
                type: integer
                description: The witness index.
          description: The commitment tree bytes.
        block_height:
          type: integer
          description: The block height of the commitment tree